refactor(analyzer): extract CFG overlay helper into call-graph-cfg.ts (slice 3d)
Next safe slice of modularize-call-graph-builder. Moves the CFG / DATA-FLOW OVERLAY
HELPER section (buildCfgFor) out of call-graph.ts into call-graph-cfg.ts — a pure,
fail-soft wrapper around buildFunctionCfg (./cfg.js) that body-resolves a declaration
wrapper (const-arrow, decorated-Python def) before building the per-function CFG
overlay.

File-internal (referenced only in test COMMENTS, never imported), so imported back,
NOT re-exported — surface unchanged. Dropped the now-unused buildFunctionCfg import
from call-graph.ts (FunctionCfg/CfgNode still used). call-graph.ts: 4,879 → 4,846 lines.

Behavior-preserving, verified four ways:
- Snapshot oracle FIRST extended to serialize the full per-function `cfgs` overlay
  (blocks/edges/defUse/params) and to exercise buildCfgFor's body-digging via an
  arrow-assigned-to-const and a decorated Python def; then captured before/after:
  IDENTICAL — SHA-256 34c7bce5…
- Public export surface byte-identical (0 added, 0 removed).
- tsc compiles all importers clean; build/lint/typecheck green.
- Full suite green: 279 files / 5534 tests.

call-graph.ts is now 5,425 → 4,846 lines across five extracted sibling modules
(types, extract, external, complexity, cfg). No feature, no dependency, no LLM.
